+++ CHANGELOG.md
@@ -23,6 +23,11 @@ Remember to align the itemized text with the first line of an item within a list
     `os.environ['JAX_JIT_PJIT_API_MERGE'] = '0'`.
     The merge must be disabled via an environment variable since it affects JAX
     at import time so it needs to be disabled before jax is imported.
+  * `axis_resources` argument of `with_sharding_constraint` is deprecated.
+    Please use `shardings` instead. There is no change needed if you were using
+    `axis_resources` as an arg. If you were using it as a kwarg, then please
+    use `shardings` instead. `axis_resources` will be removed after 3 months
+    from Feb 13, 2023.
 
 ## jaxlib 0.4.4
   * Breaking changes